Tighten operator snapshot sync and cleanup
