+++ crates/defra-agent-desktop/src/operator/mod.rs
@@ -1,3 +1,23 @@
 pub mod actions;
 pub mod controller;
 pub mod projection;
+
+mod deployments;
+mod documents;
+mod rows;
+mod shared;
+
+pub(crate) use deployments::{build_deployment_entries, DeploymentEntry};
+pub(crate) use documents::{
+    draft_for_selection, draft_matches_selection, entity_summaries, filter_entity_summaries,
+    new_draft_for_section, section_meta, EntitySummary,
+};
+pub(crate) use rows::{
+    backend_row, behavior_row, inference_profile_row, scheduled_task_row, tool_selection_row,
+};
+pub(crate) use shared::{
+    abbreviate_identifier, bool_word, compact_timestamp, normalize_optional_owned,
+    normalize_required, parse_optional_f64, parse_optional_i64, parse_optional_rfc3339,
+    parse_required_positive_i64, scheduled_task_is_due, scheduled_task_next_run_label, split_csv,
+    summarize_request_content, truncate_line,
+};
