+++ crates/defra-agent-desktop/src/views/operator/prepare.rs
@@ -1,14 +1,9 @@
 use crate::client::{ClientCore, ClientStore};
-use crate::operator::controller as operator_controller;
 use crate::state::ShellState;
 
 pub(super) fn prepare_state(
-    state: &mut ShellState,
-    client: Option<&ClientCore>,
-    store: Option<&ClientStore>,
+    _state: &mut ShellState,
+    _client: Option<&ClientCore>,
+    _store: Option<&ClientStore>,
 ) {
-    if let Some(store) = store {
-        let peer_statuses = client.map(ClientCore::peer_statuses).unwrap_or_default();
-        operator_controller::sync_from_snapshot(&mut state.operator, &peer_statuses, store);
-    }
 }
