+++ crates/defra-agent-desktop/src/views/operator/recent_failures.rs
@@ -7,84 +7,6 @@ use crate::views;
 
 use super::editors::{editor_heading, read_only_field, read_only_multiline};
 use super::request_timeline::render_request_detail;
-use super::shared::{normalize_optional_owned, truncate_line};
-use super::EntitySummary;
-
-pub(super) fn recent_failure_summaries(
-    store: &ClientStore,
-    selected_agent_did: Option<&str>,
-) -> Vec<EntitySummary> {
-    let mut rows: Vec<(Option<String>, EntitySummary)> = Vec::new();
-
-    for request in store
-        .requests
-        .iter()
-        .filter(|row| row.agent_did.as_deref() == selected_agent_did)
-    {
-        let failure = normalize_optional_owned(request.failure_reason.as_deref().unwrap_or(""))
-            .or_else(|| {
-                store
-                    .latest_response_for_request(&request.request_id)
-                    .and_then(|response| {
-                        normalize_optional_owned(response.error_message.as_deref().unwrap_or(""))
-                    })
-            });
-        let Some(failure) = failure else {
-            continue;
-        };
-
-        rows.push((
-            request
-                .claimed_at
-                .clone()
-                .or_else(|| request.created_at.clone()),
-            EntitySummary {
-                id: format!("request:{}", request.request_id),
-                title: super::shared::summarize_request_content(
-                    request.content.as_deref().unwrap_or_default(),
-                    &request.request_id,
-                ),
-                meta: format!(
-                    "request  {}  {}",
-                    request.lifecycle_state.as_deref().unwrap_or("failed"),
-                    truncate_line(&failure, 64),
-                ),
-            },
-        ));
-    }
-
-    for task in store
-        .scheduled_tasks
-        .iter()
-        .filter(|row| row.agent_did.as_deref() == selected_agent_did)
-    {
-        let Some(error) = normalize_optional_owned(task.last_error.as_deref().unwrap_or("")) else {
-            continue;
-        };
-
-        rows.push((
-            task.last_run_at.clone().or_else(|| task.updated_at.clone()),
-            EntitySummary {
-                id: format!("task:{}", task.task_id),
-                title: task.name.clone().unwrap_or_else(|| task.task_id.clone()),
-                meta: format!(
-                    "scheduled task  {}  {}",
-                    task.last_status.as_deref().unwrap_or("error"),
-                    truncate_line(&error, 64),
-                ),
-            },
-        ));
-    }
-
-    rows.sort_by(|left, right| {
-        right
-            .0
-            .cmp(&left.0)
-            .then_with(|| right.1.id.cmp(&left.1.id))
-    });
-    rows.into_iter().map(|(_, summary)| summary).collect()
-}
-
 pub(super) fn render_recent_failure_detail(ui: &mut Ui, state: &ShellState, store: &ClientStore) {
     editor_heading(ui, "Failure Detail");
     let Some(selected_id) = state.operator.selected_entity_id.as_deref() else {
