+++ crates/defra-agent-desktop/src/views/operator/request_timeline.rs
@@ -6,54 +6,6 @@ use crate::state::ShellState;
 use crate::views;
 
 use super::editors::{editor_heading, read_only_field, read_only_multiline};
-use super::shared::{abbreviate_identifier, compact_timestamp, summarize_request_content};
-use super::EntitySummary;
-
-pub(super) fn request_timeline_summaries(
-    store: &ClientStore,
-    selected_agent_did: Option<&str>,
-) -> Vec<EntitySummary> {
-    let mut rows: Vec<_> = store
-        .requests
-        .iter()
-        .filter(|row| row.agent_did.as_deref() == selected_agent_did)
-        .collect();
-    rows.sort_by(|left, right| {
-        right
-            .created_at
-            .cmp(&left.created_at)
-            .then_with(|| right.request_id.cmp(&left.request_id))
-    });
-
-    rows.into_iter()
-        .map(|row| {
-            let latest_response = store.latest_response_for_request(&row.request_id);
-            let response_state = latest_response
-                .and_then(|response| response.status.as_deref())
-                .unwrap_or("waiting");
-            EntitySummary {
-                id: row.request_id.clone(),
-                title: summarize_request_content(
-                    row.content.as_deref().unwrap_or_default(),
-                    &row.request_id,
-                ),
-                meta: format!(
-                    "{}  rsp {}  session {}  {}",
-                    row.lifecycle_state.as_deref().unwrap_or("pending"),
-                    response_state,
-                    abbreviate_identifier(row.session_id.as_deref().unwrap_or("none")),
-                    compact_timestamp(
-                        row.claimed_at
-                            .as_deref()
-                            .or(row.created_at.as_deref())
-                            .unwrap_or(""),
-                    ),
-                ),
-            }
-        })
-        .collect()
-}
-
 pub(super) fn render_request_timeline_detail(ui: &mut Ui, state: &ShellState, store: &ClientStore) {
     editor_heading(ui, "Request Detail");
     let Some(request_id) = state.operator.selected_entity_id.as_deref() else {
