+++ crates/defra-agent-desktop/tests/operator_view.rs
@@ -1,13 +1,10 @@
 use anyhow::Result;
 use chrono::Utc;
 use defra_agent_desktop::client::{
-    ClientCore, ClientCoreOptions, ClientStore, ClientStoreRows, DesktopPaths,
+    ClientCore, ClientCoreOptions, DesktopPaths,
 };
-use defra_agent_desktop::state::{OperatorDraft, OperatorSection, ShellState};
-use defra_agent_desktop::views::operator::prepare_state;
 use defra_agent_protocol::row::{
-    AgentBehaviorRow, AgentPrincipalRow, InferenceBackendRow, InferenceProfileRow,
-    ScheduledTaskRow, ToolSelectionRow,
+    AgentBehaviorRow, InferenceBackendRow, InferenceProfileRow, ScheduledTaskRow, ToolSelectionRow,
 };
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
@@ -171,79 +168,4 @@ async fn operator_document_saves_refresh_store() -> Result<()> {
 
     core.shutdown().await?;
     Ok(())
-}
-
-#[test]
-fn operator_prepare_state_reselects_entity_and_refreshes_draft() {
-    let store = ClientStore::from_rows(ClientStoreRows {
-        agent_principals: vec![AgentPrincipalRow {
-            agent_did: "did:defra:amy".to_string(),
-            display_name: Some("Amy".to_string()),
-            default_behavior_id: Some("amy-default".to_string()),
-            enabled: Some(true),
-            created_at: None,
-            created_by: None,
-        }],
-        behaviors: vec![AgentBehaviorRow {
-            behavior_id: "amy-default".to_string(),
-            agent_did: Some("did:defra:amy".to_string()),
-            display_name: Some("Amy Default".to_string()),
-            system_prompt: Some("You are Amy.".to_string()),
-            backend_id: Some("backend-amy".to_string()),
-            model_name: Some("openai/gpt-5.4".to_string()),
-            tool_selection_id: Some("tools-amy".to_string()),
-            inference_profile_id: Some("profile-amy".to_string()),
-            compaction_strategy: Some("rolling-summary".to_string()),
-            compaction_threshold: Some(0.7),
-            enabled: Some(true),
-            created_at: Some("2026-04-14T00:00:00Z".to_string()),
-        }],
-        ..ClientStoreRows::default()
-    });
-    let mut state = ShellState::default();
-    state.operator.selected_section = OperatorSection::Behaviors;
-    state.operator.selected_peer_id = Some("peer-missing".to_string());
-    state.operator.selected_agent_did = Some("did:defra:amy".to_string());
-    state.operator.selected_entity_id = Some("behavior-missing".to_string());
-    state.operator.last_apply_error = Some("stale validation".to_string());
-    state.operator.draft = Some(OperatorDraft::Behavior(
-        defra_agent_desktop::state::BehaviorDraft {
-            behavior_id: "behavior-missing".to_string(),
-            agent_did: "did:defra:amy".to_string(),
-            display_name: "Stale".to_string(),
-            system_prompt: String::new(),
-            backend_id: String::new(),
-            model_name: String::new(),
-            tool_selection_id: String::new(),
-            inference_profile_id: String::new(),
-            compaction_strategy: String::new(),
-            compaction_threshold: String::new(),
-            enabled: false,
-            created_at: String::new(),
-        },
-    ));
-
-    prepare_state(&mut state, None, Some(&store));
-
-    assert_eq!(state.operator.selected_peer_id, None);
-    assert_eq!(
-        state.operator.selected_agent_did.as_deref(),
-        Some("did:defra:amy")
-    );
-    assert_eq!(
-        state.operator.selected_entity_id.as_deref(),
-        Some("amy-default")
-    );
-    assert_eq!(
-        state.operator.last_apply_error.as_deref(),
-        Some("stale validation")
-    );
-    match state.operator.draft.as_ref() {
-        Some(OperatorDraft::Behavior(draft)) => {
-            assert_eq!(draft.behavior_id, "amy-default");
-            assert_eq!(draft.display_name, "Amy Default");
-            assert_eq!(draft.backend_id, "backend-amy");
-        }
-        other => panic!("expected refreshed behavior draft, got {other:?}"),
-    }
 }
